+++ arcgishub/utils.py
@@ -1,0 +1,13 @@
+
+def _lazy_property(fn):
+    '''Decorator that makes a property lazy-evaluated.
+    '''
+    # http://stevenloria.com/lazy-evaluated-properties-in-python/
+    attr_name = '_lazy_' + fn.__name__
+
+    @property
+    def _lazy_property(self):
+        if not hasattr(self, attr_name):
+            setattr(self, attr_name, fn(self))
+        return getattr(self, attr_name)
+    return _lazy_property
